Repared Bullshit with git (\o/)
Soo :
* Created test folder
* Created Config file
* Created MongoDB connector
* Updated development compose file adding mongodb
* Making it work like before ...
