+++ backend/requirements-dev.txt
@@ -1,4 +1,5 @@
 Flask
 Flask-Testing
+Flask-pymongo
 
 docker
