+++ stack-dev-app.yml
@@ -1,14 +1,34 @@
 version: "3.3"
 
+networks:
+   radiobretzel:
+      driver: bridge
+
 services:
-  backend:
-    image: radiobretzel/backend:feature-1.1-Generate-audio-source
-    command: app/app.py
-    build:
-      context: ./app/
-      dockerfile: dockerfiles/dev
-    ports:
-      - 80:5000
-    volumes:
-      - ./app/:/usr/src/app
-      - /var/run/docker.sock:/var/run/docker.sock
+   backend:
+      image: radiobretzel/backend:dev
+      command: run.py
+      build:
+         context: ./backend/
+         dockerfile: dockerfiles/dev
+      depends_on:
+         - database
+      ports:
+         - 80:5000
+      volumes:
+         - ./backend/:/usr/src/app
+         - /var/run/docker.sock:/var/run/docker.sock
+      networks:
+         radiobretzel:
+            aliases:
+               - backend
+               - backend.radiobretzel
+   database:
+      image: mongo
+      ports:
+         - 27017:27017
+      networks:
+         radiobretzel:
+            aliases:
+               - database
+               - database.radiobretzel
